Channel selection
Added ability to select channels that should be included for power calculations. This will be very important if there are any bad channels that need to be excluded from the analysis or if there is a particular region of interest to monitor.

Modified the axes limits for scatter plot and histogram to be Median +/- 3* Median Absolute Deviation to avoid outliers making plots unusable.