Major Update: Tabs, Watchlist
Implemented tabs in GUIDE so allow more changable settings to fit in GUI. Added a watchlist where each item is a group of channels at a particular frequency range that you want to monitor. This is to allow the user to monitor multiple regions and frequency bands during a single session.